Add yq installation and refactor deployment script usage
An installation step for yq tool has been added to the action.yaml file. The update-values-tag.sh script has also been refactored, replacing the use of 'sed' command with the yq command for updating the image tag in the deployment script. The changes aim to streamline the process and improve the robustness of YAML parsing.